refactor operator

+++ src/operators/tensor/implementations/tensor_bool.cairo
@@ -66,7 +66,7 @@ impl BoolTensor of TensorTrait<bool> {
 
     fn reduce_sum(
         self: @Tensor<bool>,
-        axes: Option<Span<usize>>,
+        axes: Option<Span<i32>>,
         keepdims: Option<bool>,
         noop_with_empty_axes: Option<bool>
     ) -> Tensor<bool> {
