+++ src/operators/tensor/implementations/tensor_fp8x23wide.cairo
@@ -77,7 +77,7 @@ impl FP8x23WTensor of TensorTrait<FP8x23W> {
 
     fn reduce_sum(
         self: @Tensor<FP8x23W>,
-        axes: Option<Span<usize>>,
+        axes: Option<Span<i32>>,
         keepdims: Option<bool>,
         noop_with_empty_axes: Option<bool>
     ) -> Tensor<FP8x23W> {
